Restructure /applications API responses
This re-formats the response to be nested JSON as the webhook provider
in External Secrets cannot retreive values at a top level JSON object
(possibly a bug). This also changes the type of the application id to a
string and removes permissions field, again as External Secrets cannot
handle parsing JSON where values are not of type String.