Fix `devise_zxcvbn` gem to version >=1.1, <2
We are using this gem to determine whether a password is considered
weak, according to the rules implemented in the gem at the time it was
first used in signon.

In later releases, these rules have changed (e.g. commonly used
passwords are not permitted). We cannot upgrade the gem as the password
rules will change but won't have been communicated to users.

Fixing to this specific version will mean we retain the existing rules.