Add rake task to sync ApiUser tokens with Kubernetes
This task will retrieve a config map from kubernetes which lists the
emails for ApiUsers whos authorisation tokens are to be synchronised
with kubernetes. The task will then create or update Kubernertes
Secrets for every token associated for that ApiUser. This is need to
automate secret creates of tokens which are used by other Applications
in Kubernetes.