Remove environments for k8s secret sync task
This removes the need to specify and have a list of emails per
environmnet for the kubernetes token secrets sync rake task. The emails
are largely consistent per env, and will just cause duplicated config in
the configmap. However, keeping this code in the commit incase is needed
in future.