Make arguments to DataHygiene.call explicit
Allowing varargs caused an issue when this function was called with multiple arguments in ruby 3, since it puts these arguments into an array, and puts any keyword arguments into a hash which then runs into ruby 3 issues.

It seems in any case better to just be explicit about the required arguments, as it requires less reading through the code to work out how to use the 'call' function.