Add task to promote all non-GDS managing editors to 'organisation admin' role
As part of turning on 2 step verification for all users, we would like any managing editors to have organisation admin roles, in order to ensure that the appropriate people in an organisation can manage their users.

We may remove this in the future, but given the possibility of these permissions becoming out of sync, it may be useful to keep.