Do not display 2sv set up banner for exempted users
If a user is exempted from 2sv, we do not want to prominently display this call to action to set up 2sv banner. However, we do want to retain a way for them to set up 2sv on their own if their circumstances change, so the smaller 'set up 2-step verification' link on the right hand side will still be displayed.