Allow `require_2sv` to be set for all users from an organisation
This allows us to set `require_2sv` on an organisation, which will then
force all their users to use 2SV, even if it is not enabled on their
account.

Exemptions will continue to work as before, even if a department has
been mandated for 2SV.

Users without an organisation (e.g. those from local authorities) are
not affected.