Do not show 2SV checkbox for new admin/superadmin users
At the moment, we require new admin and superadmin users to have 2SV set
up on their account.

However we also show a checkbox to select whether 2SV is required for
this user or not.

The checkbox is useless, as we enable 2SV on the user's account
regardless of whether the box was checked or not.

This removes the checkbox for admin and superadmin users, which makes
the flow match the actual behaviour.