Rescue from invalid date errors
If a user inputs an invalid date (e.g. the 30th February), they should get a nice error displayed to them, and the exemption should not be created.

This also refactors the controller to DRY it up a bit, introducing a function that either returns validated parameters that we know are safe to use, or an error message to display.