Ensure that if a user is exempted they have both reason and expiry date
This adds an extra layer of protection from a user getting into an invalid state, by only having one of an exemption reason, or an exemption expiry date. This should always be the case at the moment, since both must be created in the UI, and when we remove an exemption, both of these fields are wiped, but validating this at the model level ensures this stays true in the future.

This required changing an existing test that wasn't using the two_step_exempted test factory, and so didn't have an exemption date.