Add rake task to enable 2SV on an organisation
We currently have a rake task to enable 2SV for all existing users from
an organisation, however we have to use the UI to enable it on an
organisation.

This adds a rake task so it can be done without lots of clicking (e.g.
when updating organisations in bulk).