Only display API user permissions if there are any
User#permissions_for always returns an array and so the if statement
condition was always true. Looking back at this commit [1], the
equivalent if condition was checking for a similar array being empty
which makes more sense.

Although the new implementation calls User#permissions_for twice for
each iteration of the loop, the query result will be cached by
ActiveRecord.

In practice this may not make much difference, because it looks as if a
user that has access to an application always has the "signon"
permission anyway.

Co-authored-by: Chris Lowis <chris.lowis@digital.cabinet-office.gov.uk>

[1]: 5a72bc5f6d2c7fb2cdf86ebb94835834a2c999ac