Use design system for two_step_verification/show view
This commit adopts the GDS design system for the show and update
actions on the TwoStepVerificationController. A user sees this page
when they add or change the two-step verification code for their
account.

The main change to the page is the switch from collapsable sections to
having everything visible by default. We considered using the
step-by-step, accordion or tabs components/patterns but it's not
obvious that these are suitable for forms. Also since there's not a
lot of content in each "step", it's not obvious that there's a lot of
benefit to the user of hiding some of the "steps". So we've just made
the "steps" into numbered headings with non-collapsing content. We don't
think this looks too bad. If there's too much content on the page, we
could consider splitting the page into multiple pages.

We've made use of the details component to hide the explanation for
those users that can't use the barcode. The Design System recommends
this component for content that's not relevant to all users which seems
appropriate here.

As the design system recommends [1], we now show an error summary at the
top of the page with a link that takes the user to the heading for step
3 as well as highlighting the form field where the error has occurred.

The new version of the page doesn't include any of the custom analytics
tracking that was in the orginal version. We've recently discovered that
no events have been recorded in GA for this app since June 2017, so we
assume something has inadvertently broken the code. This means it
doesn't seem worthwhile trying to reimplement the custom analytics
tracking in the new version of the page. Note that we will still see
page views for this page as we do for all pages.

[1]: https://design-system.service.gov.uk/components/error-summary/

Co-authored-by: James Mead <james@floehopper.org>