Target text that user can see in 2sv integration tests
This changes the tests of the two step verification flow to click on
buttons based on the label the user can see, rather than the `name`
attribute in the URL. It makes them slightly less fragile and allows
us to redesign the view using the design system without having to
touch the tests.

We've pulled out the click_button call in the enter_2sv_code helper
into the individual tests as it feels like a button click doesn't
belong in this method.

Co-authored-by: James Mead <james@floehopper.org>