Move otp_secret_key_uri method from controller to helper
It's not idiomatic Rails to have a public method on a controller which
is not an action. I considered just making the method private, but since
it was being called from a bunch of tests I decided to move it to a
helper.

While the current_user method is probably accessible from
TwoStepVerificationHelper methods, I decided it was helpful to make the
user an explicit method argument to highlight that the URI is different
for different users.