Add resolution for stylelint packages
Due to a conflict in packages that are used in stylelint and
jasmine-browser-runner, we need to add a resolution for stylelint
packages. This ensures that stylelint uses the CJS versions of
these packages instead of the ESM versions (which it is not compatible
with). See more here alphagov/government-frontend#2781

c.f. alphagov/whitehall#7636