Set width for code input field
Unfortunately the GOV.UK design system only seems to support widths of 5
or 10 and not 6 which is actually what we want here. However, since it
seems to just set the max-width property on the input element rather
than setting the width property to an explicit number of characters, I
think setting width to 10 is appropriate here.

See this documentation [1] for more details.

[1]: https://design-system.service.gov.uk/components/text-input/#fixed-width-inputs