Fix refute_dimension_is_set assertion
The AnalyticsHelpers#assert_dimension_is_set method was changed in this
commit [1] when code to set the Google Analytics dimension8 value was
added to the admin_layout layout. The related #refute_dimension_is_set
method should have been changed at the same time, but presumably since
no tests failed it was missed. Thus the #refute_dimension_is_set
assertion has been effectively useless since then.

The only tests using #refute_dimension_is_set are
test/integration/sign_in_test.rb & test/integration/sign_out_test.rb
which relate to the SessionsController. The layout of this controller
was changed to the admin_layout layout in an even earlier commit [2].
This means that these assertions only need to work with the admin_layout
layout and not with the application layout which uses
govuk_admin_template and thus the GOVUKAdmin.setDimension function.

In order to make the #refute_dimension_is_set assertion useful again,
this commit extracts #dimension_set_js_code so that both the assert &
refute versions of the assertion are checking for the same string. Doing
this should also make it less likely that we'll make a similar mistake
in the future.

[1]: 7fe1e64b593e01e3cc1d38049c78778262a47a3d
[2]: 5173710f3f91fe35a26944b5ed3ca52a5cbeb938