Remove broken analytics event tracking code
While working on #2174 we discovered that no events have been sent to
Google Analytics from the signon app since June 2017. This strongly
suggests that nobody has been looking at them and so it seems sensible
to simplify the app by removing the relevant code.

I haven't spent any time investigating why the code hasn't been working,
but my guess is that something changed in govuk_admin_template and
the signon app wasn't updated accordingly, or perhaps a bit like
in #2190 it got broken inadvertently when some of the pages were moved
to follow the Design System.