Replace QR Code PNG with SVG
This replaces the PNG QR code on the 2SV setup page with an SVG. This
has the following benefits:

1. It removes the use of a `data:` URL which was throwing a console
error due to the content security policy provided by the
`govuk_app_config` gem[1]. This error doesn't currently cause an issue
in production as we have CSP set to "report-only" mode. We intend to
move to "enforcement" mode[2] though, and this is a step in that direction.
2. It allows the QR code to be displayed in the development
environment (the PNG image `data` URL was previously blocked due to
1.)
3. It makes the QR code responsive to the screen width

I've enabled the `use_path` option in the call to `as_svg` to reduce
the size of the generated SVG code. The `viewbox` option makes the SVG
responsive to its container width.

I've added a little padding to the left and right of the QR code's
container to ensure it doesn't bleed right to the edge of the screen
on narrow width devices. We thought this could make it harder for
someone to scan it correctly.

[1] https://github.com/alphagov/govuk_app_config/blob/main/lib/govuk_app_config/govuk_content_security_policy.rb#L46
[2]
https://docs.publishing.service.gov.uk/manual/content-security-policy.html#gov-uk-csp-history