Remove memoization of SSOPushCredential.user
SSOPushCredential.credentials is only ever called from
SSOPushClient#initialize which in turn is only ever called from
PermissionUpdater#perform & ReauthEnforcer#perform. Both of these are
subclasses of PushUserUpdatesJob which is a subclass of ApplicationJob,
i.e. they are both background jobs.

Since the call to User.find_by in SSOPushCredential is always executed
in the context of a background job, it shouldn't matter if the query is
a bit slow. And in any case there's an index on users.email in the
database and the query is very simple, so it should actually be pretty
fast. And if it's called multiple times within a single process, the
ActiveRecord cache should kick in.

A class instance variable is a bit of a dangerous thing to use, because
the state associated with it lives for as long as the *process* lives.
So I think the potential dangers and the extra code complexity isn't
warranted by any concerns about performance.