Add foreign key to oauth_access_tokens.application_id
Trello: https://trello.com/c/chkpC1am

We've been seeing ActionView::Template::Error exceptions [1] in
integration when editing the SSO Push user in integration. It appears to
be happening because that user has a couple of access tokens associated
with it that have application_id which point to non-existent
applications.

I think we could argue whether or not we actually need (or indeed
should) to be able to edit this "singleton" ApiUser, but either way it
doesn't seem good from a database integrity point of view that an access
token can end up with an application_id that points to a non-existent
application.

In the migration we create dummy applications for those access tokens to
point to. And then we add the foreign key constraint. This was the least
controversial way I could come up with to prevent the exception from
happening again.

[1]: https://govuk.sentry.io/issues/3975889554/?project=202251