Extract authenticate_access method in UserTest
Using an intention-revealing method name instead of a comment seems like
an improvement and I'm hoping to re-use the new method in a subsequent
commit.

Also it seems as if we don't need to explicitly set
Doorkeeper::AccessToken#token because it gets set by a callback in
Doorkeeper which means we can make the code a bit simpler.