Introduce a new context into TwoStepVerificationTest
I'm planning to introduce a new context at the same level as this new
one in a subsequent commit. Making this change first will make it easier
to see what changes in that commit.

The diff is best viewed using `git diff --ignore-all-space` otherwise
it's hard to see the woods for the trees with all the indentation
changes.