Don't ask for 2SV code unless user has 2SV setup
Trello: https://trello.com/c/YxXfT38U

We've been seeing NoMethodError exceptions on a POST request to
/users/two_step_verification/session. The exception was happening deep
in the bowels of the ROTP gem, but the reason why it was happening was
that `User#otp_secret_key` was `nil`.

I've managed to reproduce the problem with a user that's signed in but
doesn't have 2SV setup, by visiting the
/users/two_step_verification/session/new path (entered directly into the
browser location) and submitting a code. Since the user doesn't have 2SV
setup, `User#otp_secret_key` is `nil` and the exception is raised in
`Devise::TwoStepVerificationSessionController#create`.

I haven't managed to track down a place where the application might send
a user in this state to this page, so that remains a bit of a mystery.
However, in any case it makes sense to add a belt-and-braces before
filter to redirect the user to the home page where there will be the
option to setup 2SV if they want to.