Convert batch invitation show page to design system
This commit converts the batch invitations show page to use the design
system. I've made the following choices during this conversion:

1. Use the `notice` component to indicate when a batch is still being
processed (the previous `alert-info` div)
2. Use a `success_alert` component to indicate that all records were
successfully processed (the previous `alert-success` div)
3. Use an `error_alert` component to indicate if one or more records
in the batch couldn't be processed (the previous `alert-danger`
div). The previous alert didn't provide a mechanism for the user to
fix the error so for now I've retained that behaviour by using an
`error_alert` rather than a `form_error_summary`[1] for example.
4. I've removed the button to link to "All users". The docs suggest[2]
that buttons should only be used to "move through a transaction" and
that isn't really what it was doing here. I intend to add a
breadcrumb trail which includes "all users" in a subsequent commit to
reinstate the behaviour provided by this button.

[1] https://components.publishing.service.gov.uk/component-guide/error_summary
[2]
https://components.publishing.service.gov.uk/component-guide/button