Add a breadcrumb trail to batch invitations show page
In the previous commit I removed an "All users" button which
previously allowed a user to return to the users_path from this
page. This commit adds a breadcrumb trail that includes the previous
page, as well as the "all users" page and the home page.

This is the first time we've used a breadcrumb trail in this app - I
suspect there's more pages we'll discover that could benefit from this
pattern. For now, I've added a condition in the admin layout that uses
either the content for `back_link` or `breadcrumb` if either are
defined (prefering `back_link` in the, as yet unused, case where both
are defined).