Log exceptions when inviting batches of users
We've recently received notice that a user attempted to upload a batch
of users and some or all of those users failed. We don't currently
provide much guidance to the user on why their batches have failed,
and in fact, we don't receive any information as the product team
because those exceptions are currently being rescued silently.

This commit logs (via GovukError) some information about why the
process failed, and I'm hoping we'll be able to use that information
to improve the experience for the users, or at least be able to deal
with their issues manually via the second-line process.