Migrate api_users new page to design system
This commit moves the api_users new page to use the design system. The
changes here should be fairly uncontroversial, but have involved
rewriting most of the template.

I've specified that the `create` action should use the `admin_layout`
in the ApiUsersController so that the correct layout is used when the
page re-renders on validation error.

The `_form_fields` partial referenced in the old template is still
used by the `edit` view, which will be migrated as a separate piece of
work.

I've had to slightly change the integration test to target the error
summary div provided by the design system, and to use a regex to match
the text rather than an exact string (since the design system
`error_summary` partial suggests adding a title).