Move bulk_grant_permission_set new page to design system
We've decided to replace this page with a simple form that allows an
admin to grant the signin permission to all users for a specific
application, since that is all the current system seems to be used
for.

This means we don't have to worry about finding a design-system
alternative for the permussions multi-select box.

We think this can be extended to include specific permissions at a
later date (perhaps as a second "step" in the flow) if we want/need
to.

This commit has a working implementation that re-uses the existing
BulkGrantPermissionSet model but only passes in the signin permission
for the selected app.

This commit is the minimal change I could come up with that supports
the new behaviour. I've removed a redundant test of the error message
in bulk_granting_permisions_test but otherwise tried to keep the
changes to that test as small as possible to demonstrate that the
behaviour from before this change is retained (albeit only for the
signin permission). I'll refactor/simplify the tests and application
code now the feature is simpler in subsequent commits.