Remove Doorkeeper token_info route
This removes the following route:

```
oauth_token_info GET /oauth/token/info(.:format) doorkeeper/token_info#show
```

This endpoint is the redirect for native apps in the Implicit Grant
flow[1]. It was previously used in our
signin_required_authorizations_controller_test but I've updated that
test so that it no longer uses the Implicit Grant flow (see commit
titled "Remove test for Implicit flow and native url").

I've checked Logit and there have been no requests for this endpoint in
the 1 or 2 weeks of logs that we retain.

[1]: https://github.com/doorkeeper-gem/doorkeeper/blob/v5.6.6/lib/doorkeeper/oauth/authorization/token.rb#L77-L83