Add client side validation to catch common 2sv code errors
We've had reports from our users that when they type in the code from
their 2sv application, if they include a space (which the apps often
include in their display between each group of 3 digits) we don't
accept the code and the "Sorry that code didn’t work. Please try
again." error doesn't help them resolve the issue.

This commit improves things slightly by changing the type of the input
field to "number" and including a regex pattern to match 6 consecutive
numbers. I've added some JavaScript to further customise the error
message (by default it is "Please match the format requested") to be
more helpful.

Note that we need to clear the validation message with an `oninput`
handler to clear any errors previously set by the `oninvalid` handler
if the user has already made an invalid attempt.

We think a longer-term solution to this problem is to improve the
input field (for example by having 6 separate boxes, one for each
number). I think it would also be a good idea to change the controller
code to use a model with validations so that we can do more
comprehensive validation of the code on the server side. But those are
bigger changes and this feels like a useful improvement to the current
behaviour.