Make it more explicit which applications to bulk grant permissions
Before this commit it wasn't very clear which applications were
included in the drop down list of applications to bulk grant the
signin permission to.

Previously this view relied on passing User.new to the
shared/_user_permissions partial which in turn called[1]
`applications_and_permissions` in the
UserPermissionsControllerMethods[2] helper. Since User.new returns
false when calling `.api_user?`, `#visible_applications(user)` in this
helper was falling back to using the
`:user_permission_manageable_application` scope to get the list of
applications[3]. This scope uses `current_user` instead of the passed
in user, which for both `superadmin` and `admin` is always all
applications (order by name, in the default scope).

Since we authorize the creation of BulkGrantPermissionSet only for
superadmin and admin users, the feature would only work for those
users and would always contain all applications.

This commit makes that more explicit by introducing a helper that
calls the policy_scope directly with the current user. In the tests
for this helper I'm only exercising the `superadmin` and `admin` user
roles, because in practice it's not possible for a
BulkGrantPermissionSet to be created by any other user type.

The previous version of this page also struck through any retired
applications, I intend to restore that behaviour by removing those
applications from the list in a subsequent commit.

[1]
https://github.com/alphagov/signon/blob/convert-bulk-grant-permission-sets-new-to-design-system/app/views/shared/_user_permissions.html.erb#L15
[2] https://github.com/alphagov/signon/blob/convert-bulk-grant-permission-sets-new-to-design-system/lib/user_permissions_controller_methods.rb#L18
[3]
https://github.com/alphagov/signon/blob/convert-bulk-grant-permission-sets-new-to-design-system/app/policies/user_permission_manageable_application_policy.rb#L21