Check for signin permission in SigninPermissionGranter
This `next` statement seems to be an optimization to avoid granting
"signin" permission to users that already have it. However, previously
the condition was checking whether the user had *any* permission for the
relevant application; while the test setup was explicitly using the
"signin" permission.

I suspect that in practice if a user has *any* permissions for an
application it will always have "signin" permission, so there's unlikely
to be any practical difference. However, I think the code is now easier
to read and reasona about.

Note that in any case there's no harm (other than an performance hit) in
granting "signin" permission for a user that already has it, because
User#grant_permission is idempotent. Also SigninPermissionGranter#call
only seems to be called from rake tasks or tests, so any performance hit
is unlikely to be a problem.