Remove reference to sign-on-o-tron from basic auth realm
This app was renamed from Sign-on-o-tron to Signon some years ago, but
this Basic HTTP Authentication realm was not renamed at the same time.

I'm not sure who/what uses Basic HTTP Authentication. If it's other
apps, then I doubt they are checking the realm and so this won't cause a
problem. If it's users, then the worst that will happen is their browser
will required them to reauthenticate, because the realm has changed.