Remove reference to Signonotron from SSO Push user
This app was renamed from Sign-on-o-tron to Signon some years ago, but
this user was not renamed at the same time.

I'm pretty sure this SSO Push user name is effectively an explanatory
comment and it seems very unlikely that any client apps will be relying
on its value. In particular the gds-sso gem does *not* rely on its
value; rather it relies on the user having the "user_update_permission"
permission.