Improve consistency between two places where we ask for 2sv code
We currently ask the user for their 2sv code in two places. One place
is when they sign in[1] (if the `_signonotron2_session` cookie has
expired) the other is if they change or reset their code[2].

I'd like to create a helper to generate this input so that when we
make changes to how it works (e.g. adding validation messages or
changing hints) we can make that change in one place. To do that it
helps to first remove some differences between the two uses.

This commit reduces the duplication by:
1. using the same hint and label for both. I decide to use the "reset"
version of the hint and label, but once it's in once place I can ask
for advice on whether it should change.
2. I made both forms use the type, autocomplete and inputmode
attributes recommended in the "Confirm a
phone number" pattern[3].
3. I set the width to 10 in both cases

[1] /users/two_step_verification/session/new
[2] /users/two_step_verification
[3] https://design-system.service.gov.uk/patterns/confirm-a-phone-number/