Allow users to include spaces or dashes in their OTP codes
We've had reports from our users that when they type in the code from
their 2sv application, if they include a space (which the apps often
include in their display between each group of 3 digits) we don't
accept the code and the "Sorry that code didn’t work. Please try
again." error doesn't help them resolve the issue.

This commit allows the user to separate the digits of their code with
either space or `-` characters, and also include leading or trailing
spaces. So for a valid code `123456` we now accept, for example, `
123456 `, `123 456`, `123-456`, `12 34 56`.