Change batch_invitation_users to use utf8mb3 charset
Trello: https://trello.com/c/wMQ9H2US

Previously this table had a default charset of latin1 and when the
varchar columns were created no explicit charset was defined. This meant
that the columns were all using the latin1 charset.

This in turn meant that we could not accept unicode characters like "č"
when creating a BatchInvitationUser due to a
ActiveRecord::StatementInvalid Mysql2::Error with a message along the
lines of:

    Incorrect string value: '\xC4\x8D' for column 'name' at row 1

Most (but not all!) of the other tables (including the users table) are
using the utf8mb3 charset. This meant that while it was possible to
create a user with a unicode character in their name via the single user
UI, it was not possible via the batch invitation process using a CSV.

This migration changes the default charset for the
batch_invitation_users table to utf8mb3, but that only fixes things for
new columns; it doesn't retrospectively change the charset on the
existing columns. So the migration also modifies the relevant columns to
set an explicit charset. This fixes the problem described in the Trello
card as per the new test in test/models/batch_invitation_user_test.rb.

I did contemplate adding new columns (or even a new version of the
table) to make the database structure as simple as possible, but it
would've been a lot more work and there's no functional difference and
the difference isn't visible in db/schema.rb.