Use design system for BulkGrantPermissionSetsController#show
This commit replaces the markup in the
BulkGrantPermissionSetsController#show template with their design
system equivalents.

The `table` component does not support arbitrary HTML in the `text`
attribute, so I've decided to indicate retired applications with the
text `(retired)` rather than the `<del>` tag.

I've removed the button link to "all users" for now, and intend to
replace it with a breadcrumb trail in a subsequent commit.