Add breadcrumb trail to BulkGrantPermissionSetsController actions
This ties the new and show pages together with a common navigation,
and also allows the user to return to the "users" page (which was
previously available as a button on the show page).

I couldn't think of a better name than "Result". I think it works ok
in this context, given that the user will typically only visit the
show page after initiating a bulk permission grant operation from the
new page (unless they happen to visit the show page independently via
a direct link).