Tighten up error handling in #ignoring_requests_to_redirect_uri
I don't want the `#ignoring_requests_to_redirect_uri` method silently
ignoring exceptions that aren't a result of a request to the
`@app.redirect_uri`. This change ensures that if an
`ActionController::RoutingError` is raised for anything other than a
request to `@app.redirect_uri` then we'll be notified of it.