Avoid unnecessary conditional in GET /oauth/authorize
The `skip_authorization?` method[1] invokes the block[2] we pass to the
`skip_authorization` option in our Doorkeeper initializer. The block we
pass to `skip_authorization` always returns true so the `else` branch of
this controller action is never executed.

The `skip_authorization` block was set to return true in bc9c1f017 (14
Jan 2013).

[1]: https://github.com/doorkeeper-gem/doorkeeper/blob/v5.6.6/lib/doorkeeper/helpers/controller.rb#L71
[2]: https://github.com/alphagov/signon/blob/794d412c52bafe40a43973d96c3533334a2e69cd/config/initializers/doorkeeper.rb#L351-L353