Improve error messages for SuspensionsController#edit
Before this commit, if a User could not be suspended (for example if a
reason wasn't given) we displayed a flash notice "failed". This commit
improves that error handling in line with the design system by using
`error_summary` component and adding error messages to the reason
field.

I've achieved this by modeling the "suspension" as a ruby class which
mixes in ActiveRecord::Validations so we can treat those fields on
User more like a standard record.

I did consider modelling suspensions explicitly in the database and
migrating the suspension related fields to a new table before
associating them with the User. However the suspension functionality
is provided through a custom devise module in
Devise::Models::Suspendable and it all seemed a bit too difficult to
untangle simply to improve error handling.