Reinstate pagination
I temporarily removed the pagination on the event logs page in an
earlier commit to make the changes easier to follow. Now we have a
design-system kaminari theme for pagination we can re-add it to the
page.

Note the previous version of this page used the pagination component
at the top and the bottom of the page. This doesn't seem to be how it
is used in the design system examples, so I've included it once, below
the table of events.